// app/(authenticated)/(tabs)/crypto.tsx
import { View, Text, Image, ScrollView } from "react-native";
import React, { useEffect } from "react";
import { useQuery } from "@tanstack/react-query";
import { Currency } from "@/interfaces/crypto";
import { Href, Link } from "expo-router";
import { TouchableOpacity } from "react-native-gesture-handler";
import { useHeaderHeight } from "@react-navigation/elements";
import Colors from "@/constants/Colors";
import { defaultStyles } from "@/constants/Styles";
import { Ionicons } from "@expo/vector-icons";

type CryptoHref = `/crypto/${number}`;
export default function Page() {
  const headerHeight = useHeaderHeight();

  const currencies = useQuery({
    queryKey: ["listings"],
    queryFn: () => fetch("/api/listings").then((res) => res.json()),
  });

  const ids = currencies.data
    ?.map((currency: Currency) => currency.id)
    .join(",");

  const { data } = useQuery({
    queryKey: ["info", ids],
    queryFn: () => fetch(`/api/info?ids=${ids}`).then((res) => res.json()),
    enabled: !!ids,
  });

  return (
    <ScrollView
      contentContainerStyle={{ paddingTop: headerHeight }}
      style={{ backgroundColor: Colors.background }}
    >
      <Text style={defaultStyles.sectionHeader}>Top 5 Criptomoedas</Text>
      <View style={defaultStyles.block}>
        {currencies.data?.map((currency: Currency) => (
          <Link
            href={`/crypto/${currency.id}` as Href}
            key={currency.id}
            asChild
          >
            <TouchableOpacity
              style={{
                flexDirection: "row",
                gap: 14,
                alignItems: "center",
              }}
            >
              {data?.[currency.id].logo && (
                <Image
                  source={{ uri: data?.[currency.id].logo }}
                  style={{ width: 32, height: 32 }}
                />
              )}

              <View style={{ flex: 1, gap: 2 }}>
                <Text style={{ fontWeight: "600", color: Colors.dark }}>
                  {currency.name}
                </Text>
                <Text style={{ color: Colors.gray }}>{currency.symbol}</Text>
              </View>
              <View style={{ gap: 2, alignItems: "flex-end" }}>
                <Text>
                  {currency.quote.BRL.price.toLocaleString("pt-BR", {
                    style: "currency",
                    currency: "BRL",
                  })}
                </Text>
                <View style={{ flexDirection: "row", gap: 4 }}>
                  <Ionicons
                    name={
                      currency.quote.BRL.percent_change_1h > 0
                        ? "caret-up"
                        : "caret-down"
                    }
                    size={16}
                    color={
                      currency.quote.BRL.percent_change_1h > 0 ? "green" : "red"
                    }
                  />
                  <Text
                    style={{
                      color:
                        currency.quote.BRL.percent_change_1h > 0
                          ? "green"
                          : "red",
                    }}
                  >
                    {currency.quote.BRL.percent_change_1h.toFixed(2)} %
                  </Text>
                </View>
              </View>
            </TouchableOpacity>
          </Link>
        ))}
      </View>
    </ScrollView>
  );
}
